Robot (Rosie) based testing fixes
- Reversed motor polarity to make 'bot move fwd/bkwd correctly 
- Changed encoders pulses per rotation to match number of axles on 'bot,
rather than number of stripes on a disk (which didn't work).
- Made sure the teleop command is actually started so that the command
gets scheduled and the 'bot responds to stick driving.
- Reset distance display on the dashboard to assist in debugging.
- Show encoder counts on dashboard to assist in debugging
- Update stick scaling algorithm based on on-bot testing. Needed to
negate raw stick values to make the move axis (Y) have raw negative
values move forward (ie in the positive direction).
- Left linear scaling algorithm commented out and the cubic algorithm
active as this seems more usable to the drivers.