Shooter Commands & Subsystem
Not yet tested. Removed Servo as it won't be used in the assembly
Refactored code to make it more readable.
Created SubSystemInit method to encapsulate all the subsystem
instantiation. This has to be done before making a OperatorInterface, as
the latter depends on the objects created by the former
Made PlaceGear the default Autonomous mode command
Made the winch go for 1s (rather than 4s) after the limit switch trips
on advice from Art.