Refactoring for encapsulation
Encapsulated subsystems so that they get created in the Robot class, and
not in OperatorInterface; rather they get passed in as dependencies.

Removed time delay between winch limit switch tripping and stopping the
motor. On advice of Art.

Changed winch button to activate while held, in an attempt to allow the
motor to be restarted. Not yet tested. Issue was that whenPressed
results in isFinished eventually returning true; which means the
winchRaise command is now dead - can't be restarted without re-starting
the robot code. This is an attempt to be able to restart the winch after
the limit switch cuts it off.